Use interval timedeltas instead of Frequency
In plugin class definitions, an 'interval' attribute is now the means to specify
the desired frequency of cache cycles. Frequency instances are now created all
at once and synchronized by the statscache producer factory. The associated
frequency instance is now provided to the __init__() method of plugin classes
and is made available as a 'frequency' instance attribute by default.